requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2
qdrant-client==1.11.3
openpyxl==3.1.5
